meshes: fix the core of pyplanes to decouple GmshMesh from the problem construction routines
