+++ pyPLANES/core/fem_problem.py
@@ -41,9 +41,11 @@ from scipy.sparse import coo_matrix, csc_matrix, csr_matrix, linalg as sla
 from pyPLANES.fem.fem_preprocess import fem_preprocess
 from pyPLANES.utils.io import plot_fem_solution, export_paraview
 
+from pyPLANES.gmsh.mesh import GmshMesh
+
 from mediapack import Air
 
-class FemProblem(Mesh, Calculus):
+class FemBase(Mesh, Calculus):
     def __init__(self, **kwargs):
         Calculus.__init__(self, **kwargs)
         self.result.Solver = type(self).__name__
@@ -66,7 +68,6 @@ class FemProblem(Mesh, Calculus):
         self.result.n_dof = []
         self.result.D_lambda = []
 
-
     def create_linear_system(self, omega):
         # Initialisation of the lists
         Calculus.create_linear_system(self, omega)
@@ -79,6 +80,7 @@ class FemProblem(Mesh, Calculus):
             self.update_system(*_ent.update_system(omega))
             # if _ent.dim == 2:
             #     self.update_Q(*_ent.update_Q())
+
     def update_frequency(self, omega):
             Calculus.update_frequency(self, omega)
             self.F_i, self.F_v = [], []
@@ -92,7 +94,6 @@ class FemProblem(Mesh, Calculus):
         self.Q_i.extend(_A_i)
         self.Q_j.extend(_A_j)
         self.Q_v.extend(_A_v)
-
 
     def update_system(self, _A_i, _A_j, _A_v, _F_i, _F_v, _T_i=None, _T_j=None, _T_v=None):
         self.A_i.extend(_A_i)
@@ -206,3 +207,8 @@ class FemProblem(Mesh, Calculus):
         #     mail = " mailx -s \"FEM pyPLANES Calculation of " + self.name_project + " over on \"" + self.name_server + " olivier.dazel@univ-lemans.fr < " + self.info_file.name
         #     os.system(mail)
 
+
+class FemProblem(FemBase, GmshMesh):
+    def __init__(self, **kwargs):
+        FemBase.__init__(self, **kwargs)
+        GmshMesh.__init__(self, **kwargs)
